% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGThemes-package.R
\name{DGThemes-exports}
\alias{DGThemes-exports}
\alias{\%>\%}
\title{DGThemes exported operators}
\description{
The following functions are imported and then re-exported
from the DGThemes package to enable use of the magrittr
pipe operator with no additional library calls
}
