% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGThemes-package.R
\docType{package}
\name{DGThemes}
\alias{DGThemes}
\alias{DGThemes-package}
\title{Additional Themes and Theme Components for 'ggplot2'}
\description{
A compilation of extra themes and theme components for 'ggplot2' with an
emphasis on typography.
}
\details{

}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/dungates/DGThemes}
  \item Report bugs at \url{https://github.com/dungates/DGThemes/issues}
}

}
\author{
Duncan Gates (duncangates.me)
}
\keyword{internal}
