% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-duncan.R
\docType{data}
\name{font_fs}
\alias{font_fs}
\title{Fira Sans font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_fs
}
\description{
\code{font_fn} == "\verb{Fira Sans}"
}
\keyword{datasets}
