% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-theme-duncan.R
\name{gt_theme_duncan}
\alias{gt_theme_duncan}
\title{Personalized gt theme}
\usage{
gt_theme_duncan(data, bg_color = "#FFFFFF", all_caps = F, ...)
}
\arguments{
\item{data}{A `gt_tbl` object, generated from [gt::gt].}

\item{bg_color}{The background color of the table cells.}

\item{all_caps}{Whether or not to capitalize titles}

\item{...}{Additional arguments passed to [gt::tab_options].}
}
\value{
An object of class `gt_tbl`.
}
\description{
A custom theme for tables generated with [gt::gt].
}
\examples{
mtcars \%>\%
  head() \%>\%
  gt::gt() \%>\%
  gt::tab_header(title = "Cars n stuff", subtitle = "I don't know anything about cars") \%>\%
  DGThemes::gt_theme_duncan()
}
