% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-duncan.R
\docType{data}
\name{font_os}
\alias{font_os}
\title{Open Sans font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_os
}
\description{
\code{font_an} == "\verb{Open Sans}"
}
\keyword{datasets}
