% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{premium_pal}
\alias{premium_pal}
\title{A muted, qualitative color palette}
\usage{
premium_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(premium_pal()(9))
}
