% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium-pdf.R
\name{premium_pdf}
\alias{premium_pdf}
\title{premium R markdown template for PDF output}
\usage{
premium_pdf(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Template for creating an R markdown documents with an emphasis on typography
}
