% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-theme.R
\name{theme_map}
\alias{theme_map}
\title{Mapping theme}
\usage{
theme_map(base_family = "Fira Sans", base_size = 15, bg = "gray15")
}
\arguments{
\item{base_family, }{the base family}

\item{base_size}{the base font sizer}

\item{bg}{background color}
}
\description{
theme for making maps
}
\examples{
library(sf)
library(ggplot2)
nc <- sf::read_sf(system.file("shape/nc.shp", package="sf"))
ggplot(nc) +
   geom_sf() +
   theme_map()

}
