% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-premium.r
\docType{data}
\name{font_an}
\alias{font_an}
\title{Tisa Sans Pro font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_an
}
\description{
\code{font_an} == "\verb{Tisa Sans Pro}"
}
\keyword{datasets}
