% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-premium.r
\docType{data}
\name{font_tsn}
\alias{font_tsn}
\title{Tisa Sans Pro font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_tsn
}
\description{
\code{font_tsn} == "\verb{Tisa Sans Pro}"
}
\keyword{datasets}
